third_party/blink/web_tests/svg/dynamic-updates/SVGFEConvolveMatrixElement-svgdom-divisor-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'divisor' property of the SVGFEConvolveMatrixElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS convolveMatrixElement.divisor.baseVal is 20
PASS convolveMatrixElement.divisor.baseVal is 12
PASS successfullyParsed is true

TEST COMPLETE
